+++ packages/adapter-interfaces/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @authhero/adapter-interfaces
 
+## 0.33.0
+
+### Minor Changes
+
+- add sendgrid and postmark mail services
+- migrate dbconnections and setup email providers
+
 ## 0.32.1
 
 ### Patch Changes