+++ packages/kysely/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @authhero/kysely-adapter
 
+## 0.25.3
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.33.0
+
 ## 0.25.2
 
 ### Patch Changes